refactor: move wake-on-ingress-traffic spec-read helpers to pkg/utils
Relocate WakeOnIngressTrafficEnabled and WakeOnIngressTrafficPauseTimeout
from api/v1alpha1 to pkg/utils so the API package stays a pure type
definition; gateway, controller and route projection consume the shared
helpers from pkg/utils.
