+++ pkg/utils/utils.go
@@ -385,6 +385,27 @@ func IsSandboxResumable(sbx *agentsv1alpha1.Sandbox) (bool, string) {
 	return false, "SandboxPhaseNotAllowed"
 }
 
+// WakeOnIngressTrafficEnabled reports whether the sandbox opted into
+// wake-on-traffic via its spec.
+func WakeOnIngressTrafficEnabled(sbx *agentsv1alpha1.Sandbox) bool {
+	return sbx != nil && sbx.Spec.AutoPausePolicy != nil &&
+		sbx.Spec.AutoPausePolicy.Resume != nil &&
+		sbx.Spec.AutoPausePolicy.Resume.WhenIngressTraffic != nil
+}
+
+// WakeOnIngressTrafficPauseTimeout returns the auto-pause timeout to re-arm
+// after a traffic wake, or 0 when the rule does not set a positive value.
+func WakeOnIngressTrafficPauseTimeout(sbx *agentsv1alpha1.Sandbox) time.Duration {
+	if sbx == nil || sbx.Spec.AutoPausePolicy == nil || sbx.Spec.AutoPausePolicy.Resume == nil {
+		return 0
+	}
+	rule := sbx.Spec.AutoPausePolicy.Resume.WhenIngressTraffic
+	if rule == nil || rule.PauseTimeout == nil || rule.PauseTimeout.Duration <= 0 {
+		return 0
+	}
+	return rule.PauseTimeout.Duration
+}
+
 // GetTemplateSpec resolves and returns the PodTemplateSpec from the EmbeddedSandboxTemplate.
 // If TemplateRef is specified, it will fetch the SandboxTemplate using the client.
 func GetTemplateSpec(ctx context.Context, cli client.Client, namespace string, embedded *agentsv1alpha1.EmbeddedSandboxTemplate) (*corev1.PodTemplateSpec, error) {
